# docs/swagger.yaml
openapi: 3.0.0
info:
  title: Jobsy API
  version: "1.0.0"
  description: API для регистрации, входа, подтверждения e-mail, восстановления пароля и управления профилями фрилансера и заказчика.
...
paths:
...
  /freelancers:
    get:
      tags:
        - freelancer
      summary: Просмотр списка фрилансеров
      description: Возвращает список фрилансеров с базовой информацией
      responses:
        '200':
          description: Список фрилансеров получен успешно.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FreelancerProfile'
        '404':
          description: Фрилансеры не найдены
  /freelancers/{id}:
    get:
      tags:
        - freelancer
      summary: Получить детальную информацию о фрилансере
      description: Возвращает детальную информацию о фрилансере по его ID
      parameters:
        - name: id
          in: path
          description: ID фрилансера.
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Информация о фрилансере получена успешно
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FreelancerProfile'
        '404':
          description: Фрилансер не найден
...
  /clients:
    get:
      tags:
        - client
      summary: Просмотр списка клиентов
      description: Возвращает список клиентов с базовой информацией.
      responses:
        '200':
          description: Список клиентов получен успешно.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ClientProfile'
        '404':
          description: Клиенты не найдены.
        '500':
          description: Внутренняя ошибка сервера
  /clients/{id}:
    get:
      tags:
        - client
      summary: Получить детальную информацию о клиенте
      description: Возвращает детальную информацию о клиенте по его ID
      parameters:
        - name: id
          in: path
          description: ID клиента
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Информация о клиенте получена успешно
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClientProfile'
        '404':
          description: Клиент не найден
        '500':
          description: Внутренняя ошибка сервера
...
  /projects:
    get:
      tags:
        - client
      summary: Получить список проектов
      description: >
        Возвращает список всех проектов. При передаче query-параметра "status" возвращаются проекты с указанным статусом
        Допустимые значения статуса: open (открытые), in_progress (в работе), completed (завершённые)
      parameters:
        - name: status
          in: query
          description: Фильтр по статусу проекта.
          required: false
          schema:
            type: string
            enum: [open, in_progress, completed]
      responses:
        '200':
          description: Список проектов получен успешно
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Project'
        '404':
          description: Проекты не найдены
    post:
      tags:
        - client
      summary: Создать новый проект
      description: Клиент может создать новый проект. При создании проект получает статус "open"
      security:
        - bearerAuth: []
      requestBody:
        description: Данные нового проекта
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectCreate'
      responses:
        '201':
          description: Проект успешно создан.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '400':
          description: Ошибка создания проекта
  /projects/{projectId}:
    get:
      tags:
        - client
      summary: Получить проект по id
      description: Возвращает данные одного проекта"
      security:
        - bearerAuth: []
      parameters:
        - name: projectId
          in: path
          description: ID проекта
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Проект успешно получен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '500':
          description: Внутренняя ошибка сервера
    put:
      tags:
        - client
      summary: Обновить проект
      description: Клиент может обновить проект, если его статус "open". Проекты в работе или завершённые редактировать нельзя
      security:
        - bearerAuth: []
      parameters:
        - name: projectId
          in: path
          description: ID проекта
          required: true
          schema:
            type: integer
      requestBody:
        description: Данные для обновления проекта
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectUpdate'
      responses:
        '200':
          description: Проект успешно обновлён
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '400':
          description: Проект нельзя обновить, так как его статус не "open"
    delete:
      tags:
        - client
      summary: Удалить проект
      description: Клиент может удалить проект, если его статус "open". Проекты в работе или завершённые удалять нельзя
      security:
        - bearerAuth: []
      parameters:
        - name: projectId
          in: path
          description: ID проекта для удаления.
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Проект успешно удалён.
        '400':
          description: Проект нельзя удалить, так как его статус не "open"
...
  /admin/freelancers:
    get:
      tags:
        - admin
      summary: Просмотр списка фрилансеров
      description: Возвращает список всех фрилансеров с базовой информацией
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Список фрилансеров получен успешно
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FreelancerProfile'
        '401':
          description: Нет прав
        '404':
          description: Фрилансеры не найдены
        '500':
          description: Внутренняя ошибка сервера
  /admin/freelancers/{userId}:
    get:
      tags:
        - admin
      summary: Получить детальную информацию о фрилансере
      description: Возвращает детальную информацию о фрилансере по его userId
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          description: ID фрилансера
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Информация о фрилансере получена успешно
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FreelancerProfile'
        '401':
          description: Нет прав
        '404':
          description: Фрилансер не найден
        '500':
          description: Внутренняя ошибка сервера
    delete:
      tags:
        - admin
      summary: Удалить фрилансера
      description: Удаляет профиль фрилансера по userId
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Фрилансер успешно деактивирован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefaultResponse'
        '401':
          description: Пользователь не аутентифицирован или не имеет прав
        '404':
          description: Профиль не найден
        '500':
          description: Внутренняя ошибка сервера
    put:
      tags:
        - admin
      summary: Активировать/деактивировать фрилансера
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Операция выполнена успешно
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefaultResponse'
        '401':
          description: Не аутентифицирован или нет прав
        '404':
          description: Фрилансер не найден
        '500':
          description: Внутренняя ошибка сервера
  /admin/freelancers/{userId}/activate:
    put:
      tags:
        - admin
      summary: Активировать фрилансера
      description: Меняет статус фрилансера на активный (isActive=true)
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Фрилансер успешно активирован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefaultResponse'
        '401':
          description: Не аутентифицирован или нет прав
        '404':
          description: Фрилансер не найден
        '500':
          description: Внутренняя ошибка сервера
  /admin/freelancers/{userId}/deactivate:
    put:
      tags:
        - admin
      summary: Деактивировать фрилансера
      description: Меняет статус фрилансера на неактивный (isActive=false)
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Фрилансер успешно деактивирован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefaultResponse'
        '401':
          description: Не аутентифицирован или нет прав
        '404':
          description: Фрилансер не найден
        '500':
          description: Внутренняя ошибка сервера
  /admin/clients:
    get:
      tags:
        - admin
      summary: Просмотр списка клиентов
      description: Возвращает список всех клиентов с базовой информацией
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Список клиентов получен успешно
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ClientProfile'
        '401':
          description: Нет прав
        '404':
          description: Клиенты не найдены
        '500':
          description: Внутренняя ошибка сервера
  /admin/freelancers/{userId}/projects:
    get:
      tags:
        - admin
      summary: Получить проекты, где фрилансер является исполнителем
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Проекты успешно получены
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Project'
        '401':
          description: Нет прав
        '404':
          description: Фрилансер не найден
        '500':
          description: Ошибка сервера
  /admin/clients/{userId}:
    get:
      tags:
        - admin
      summary: Получить детальную информацию о клиенте
      description: Возвращает детальную информацию о клиенте по его ID
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          description: ID клиента
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Информация о клиенте получена успешно
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClientProfile'
        '401':
          description: Нет прав
        '404':
          description: Клиент не найден
        '500':
          description: Внутренняя ошибка сервера
    delete:
      tags:
        - admin
      summary: Удалить заказчика
      description: Удаляет профиль пользователя по userId с ролью CLIENT
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Заказчик успешно удалён
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefaultResponse'
        '401':
          description: Пользователь не аутентифицирован или не имеет прав
        '404':
          description: Профиль не найден
        '500':
          description: Внутренняя ошибка сервера
  /admin/clients/{userId}/activate:
    put:
      tags:
        - admin
      summary: Активировать заказчика
      description: Меняет статус заказчика на активный (isActive=true)
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Заказчик успешно активирован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefaultResponse'
        '401':
          description: Не аутентифицирован или нет прав
        '404':
          description: Заказчик не найден
        '500':
          description: Внутренняя ошибка сервера
  /admin/clients/{userId}/deactivate:
    put:
      tags:
        - admin
      summary: Деактивировать заказчика
      description: Меняет статус заказчика на неактивный (isActive=false)
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Заказчик успешно деактивирован
          content:
            application/json:
              schema:
                 $ref: '#/components/schemas/DefaultResponse'
        '401':
          description: Не аутентифицирован или нет прав
        '404':
          description: Заказчик не найден
        '500':
          description: Внутренняя ошибка сервера
  /admin/clients/{id}/projects:
    get:
      tags:
        - admin
      summary: Просмотр проектов клиента
      description: Возвращает список всех проектов, созданных указанным клиентом
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          description: ID клиента
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Список проектов клиента получен успешно
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Project'
        '401':
          description: Нет прав
        '404':
          description: Проекты для данного клиента не найдены
        '500':
          description: Внутренняя ошибка сервера
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    DefaultResponse:
      type: object
      properties:
        message:
          type: string
          example: "Операция выполнена успешно"
...
    FreelancerProfile:
      type: object
      description: Полные данные профиля фрилансера для отображения
      properties:
        id:
          type: integer
          description: Идентификатор профиля фрилансера
          example: 1
        about:
          $ref: '#/components/schemas/FreelancerProfileAbout'
        basic:
          $ref: '#/components/schemas/FreelancerProfileBasic'
        contact:
          $ref: '#/components/schemas/FreelancerProfileContact'
        user:
          $ref: '#/components/schemas/User'
        createdAt:
          type: string
          format: date-time
          description: Дата создания профиля
          example: "2024-03-30T12:00:00"
        updatedAt:
          type: string
          format: date-time
          description: Дата последнего обновления профиля
          example: "2024-03-30T12:00:00"
        avatarUrl:
          type: string
          description: url аватарки фрилансера
        averageRating:
          type: number
          format: double
          description: Средний рейтинг фрилансера
          example: 4.5
        ratingCount:
          type: integer
          description: Количество полученных оценок
          example: 12
      required:
        - id
        - createdAt
...